+++ course/format/tiles/classes/output/course_output.php
@@ -185,7 +185,7 @@ class course_output implements \renderable, \templatable {
         // Now we can go off and get the specific data for the single or multiple page as required.
         if ($this->sectionnum !== null) {
             // We are outputting a single section page.
-            $data['showsinglesectionlegacynav'] = !$data['jsnavadminallowed'] || !$data['jsnavuserenabled'];
+            $data['showsinglesectionlegacynav'] = true;
             if ($this->sectionnum == 0) {
                 return $this->append_section_zero_data($data, $output);
             } else {
@@ -421,9 +421,11 @@ class course_output implements \renderable, \templatable {
         if (strlen('single_sec_content') > $longsectionlength) {
             $data['single_sec_content_is_long'] = true;
         }
-        $previousnext = $this->get_previous_next_section_ids($thissection->section);
-        $data['previous_tile_id'] = $previousnext['previous'];
-        $data['next_tile_id'] = $previousnext['next'];
+        if (!$this->fromajax) {
+            $previousnext = $this->get_previous_next_section_numbers($thissection->section);
+            $data['previous_tile_id'] = $previousnext['previous'];
+            $data['next_tile_id'] = $previousnext['next'];
+        }
 
         $data['visible'] = $thissection->visible;
         // If user can view hidden items, include the explanation as to why an item is hidden.
@@ -1242,31 +1244,31 @@ class course_output implements \renderable, \templatable {
     }
 
     /**
-     * For the legacy navigation arrows, establish the id of the next and previous sections
-     * @param int $currentsectionnum the id of the section we are in
-     * @return array previous and next ids
+     * For the legacy navigation arrows, establish the section number of the next and previous sections.
+     * @param int $currentsectionnum the section number of the section we are in.
+     * @return array previous and next section numbers.
      */
-    private function get_previous_next_section_ids($currentsectionnum) {
+    private function get_previous_next_section_numbers(int $currentsectionnum): array {
         $visiblesectionnums = [];
         $currentsectionarrayindex = -1;
         foreach ($this->modinfo->get_section_info_all() as $section) {
+            if ($section->section == 0) {
+                continue;
+            }
             if ($section->uservisible) {
                 $visiblesectionnums[] = $section->section;
-                if ($section->section == $currentsectionnum) {
-                    $currentsectionarrayindex = $section->section;
+                if ($section->section <= $currentsectionnum) {
+                    $currentsectionarrayindex++;
                 }
             }
         }
-        if ($currentsectionarrayindex == 0) {
-            $previous = 0; // There is no previous.
-        } else {
-            $previous = $visiblesectionnums[$currentsectionarrayindex - 1];
-        }
-        if ($currentsectionarrayindex == count($visiblesectionnums) - 1) {
-            $next = 0; // There is no next.
-        } else {
-            $next = $visiblesectionnums[$currentsectionarrayindex + 1];
-        }
+
+        // If $currentsectionarrayindex is zero, this means we are on the first available section so there is no "previous".
+        $previous = $currentsectionarrayindex == 0 ? 0 : $visiblesectionnums[$currentsectionarrayindex - 1];
+
+        // If there is no item at the next index, there is no "next" (so set next to zero).
+        $next = $visiblesectionnums[$currentsectionarrayindex + 1] ?? 0;
+
         return array('previous' => $previous, 'next' => $next);
     }