+++ mod/attendance/classes/form/studentattendance.php
@@ -47,14 +47,24 @@ class studentattendance extends \moodleform {
         $statuses = $attblock->get_statuses();
         // Check if user has access to all statuses.
         $disabledduetotime = false;
-        foreach ($statuses as $status) {
-            if ($status->studentavailability === '0') {
-                unset($statuses[$status->id]);
+        $sessionstarttime = empty($attforsession->studentsearlyopentime) ?
+            $attforsession->sessdate : $attforsession->sessdate - $attforsession->studentsearlyopentime;
+
+        if (time() < $sessionstarttime) {
+            foreach ($statuses as $status) {
+                if ($status->availablebeforesession == 0) {
+                    unset($statuses[$status->id]);
+                }
             }
-            if (!empty($status->studentavailability) &&
-                time() > $attforsession->sessdate + ($status->studentavailability * 60)) {
-                unset($statuses[$status->id]);
-                $disabledduetotime = true;
+        } else if (time() > $sessionstarttime) {
+            foreach ($statuses as $status) {
+                if ($status->studentavailability === '0') {
+                    unset($statuses[$status->id]);
+                } else if (!empty($status->studentavailability
+                    && time() > $attforsession->sessdate + ($status->studentavailability * 60))) {
+                    unset($statuses[$status->id]);
+                    $disabledduetotime = true;
+                }
             }
         }