+++ mod/attendance/preferences.php
@@ -55,7 +55,7 @@ $PAGE->set_heading($course->fullname);
 $PAGE->force_settings_menu(true);
 $PAGE->set_cacheable(true);
 $PAGE->navbar->add(get_string('settings', 'attendance'));
-
+$PAGE->requires->js_call_amd('mod_attendance/statusset', 'init');
 $errors = array();
 
 // Check sesskey if we are performing an action.
@@ -133,6 +133,8 @@ switch ($att->pageparams->action) {
         $description    = required_param_array('description', PARAM_TEXT);
         $grade          = required_param_array('grade', PARAM_RAW);
         $studentavailability = optional_param_array('studentavailability', null, PARAM_RAW);
+        $availability = optional_param_array('availability', null, PARAM_RAW);
+        $availablebeforesession = optional_param_array('availablebeforesession', '0', PARAM_RAW);
         $unmarkedstatus = optional_param('setunmarked', null, PARAM_INT);
 
         foreach ($grade as &$val) {
@@ -146,8 +148,16 @@ switch ($att->pageparams->action) {
             if ($unmarkedstatus == $id) {
                 $setunmarked = true;
             }
+            if (!isset($availablebeforesession[$id])) {
+                $availablebeforesession[$id] = 0;
+            }
+            if ($availability[$id] === '0') {
+                $studentavailability[$id] = '0';
+            } else if (empty($availability[$id])) {
+                $studentavailability[$id] = '';
+            }
             $errors[$id] = attendance_update_status($status, $acronym[$id], $description[$id], $grade[$id],
-                                                    null, $att->context, $att->cm, $studentavailability[$id], $setunmarked);
+                                                    null, $att->context, $att->cm, $studentavailability[$id], $availablebeforesession[$id], $setunmarked);
         }
         attendance_update_users_grade($att);
         break;